move WorkflowFixtures to ert.plugins
