+++ src/ert/gui/tools/plugins/plugin_runner.py
@@ -6,8 +6,7 @@ from typing import TYPE_CHECKING, Any
 
 from _ert.threading import ErtThread
 from ert.config import ErtConfig
-from ert.config.workflow_fixtures import WorkflowFixtures
-from ert.plugins import CancelPluginException
+from ert.plugins import CancelPluginException, WorkflowFixtures
 from ert.runpaths import Runpaths
 from ert.workflow_runner import WorkflowJobRunner
