+++ src/ert/plugins/ert_script.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING, Any, TypeAlias
 
 from typing_extensions import deprecated
 
-from ert.config.workflow_fixtures import WorkflowFixtures
+from .workflow_fixtures import WorkflowFixtures
 
 if TYPE_CHECKING:
     from ert.config import ErtConfig
